requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
Django==5.1
django-cors-headers==4.4.0
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.9.0
redis==5.1.0
simplejwt==2.0.1
sqlparse==0.5.1
typing==3.7.4.3
typing_extensions==4.12.2
